tests and some code fixes

+++ src/main/java/bio/terra/service/dataset/flight/create/CreateDatasetAssetStep.java
@@ -1,5 +1,6 @@
 package bio.terra.service.dataset.flight.create;
 
+import bio.terra.common.FlightUtils;
 import bio.terra.model.AssetModel;
 import bio.terra.model.ErrorModel;
 import bio.terra.service.configuration.ConfigEnum;
@@ -11,11 +12,14 @@ import bio.terra.service.dataset.DatasetJsonConversion;
 import bio.terra.service.dataset.DatasetRelationship;
 import bio.terra.service.dataset.DatasetService;
 import bio.terra.service.dataset.DatasetTable;
+import bio.terra.service.dataset.exception.InvalidAssetException;
+import bio.terra.service.dataset.flight.DatasetWorkingMapKeys;
 import bio.terra.service.job.JobMapKeys;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.FlightMap;
 import bio.terra.stairway.Step;
 import bio.terra.stairway.StepResult;
+import bio.terra.stairway.StepStatus;
 import org.springframework.http.HttpStatus;
 
 import java.util.HashMap;
@@ -73,16 +77,11 @@ public class CreateDatasetAssetStep implements Step {
         // TODO: Asset columns and tables need to match things in the dataset schema
         Dataset dataset = getDataset(context);
         FlightMap map = context.getWorkingMap();
+
         // get the dataset assets that already exist --asset name needs to be unique
         AssetSpecification newAssetSpecification = getNewAssetSpec(context, dataset);
         List<AssetSpecification> datasetAssetSpecificationList = dataset.getAssetSpecifications();
-        if (datasetAssetSpecificationList.stream()
-            .anyMatch(asset -> asset.getName().equalsIgnoreCase(newAssetSpecification.getName()))) {
-            map.put(JobMapKeys.STATUS_CODE.getKeyName(), HttpStatus.BAD_REQUEST);
-            map.put(JobMapKeys.RESPONSE.getKeyName(),
-                new ErrorModel().message("Asset already exists: " + newAssetSpecification.getName()));
-            return StepResult.getStepResultSuccess();
-        }
+
         // add a fault that forces an exception to make sure the undo works
         try {
             configService.fault(ConfigEnum.CREATE_ASSET_FAULT, () -> {
@@ -91,24 +90,35 @@ public class CreateDatasetAssetStep implements Step {
         } catch (Exception e) {
             throw new RuntimeException(e);
         }
-        assetDao.create(newAssetSpecification, dataset.getId());
+
+        try {
+            assetDao.create(newAssetSpecification, dataset.getId());
+        } catch (InvalidAssetException e) {
+            FlightUtils.setErrorResponse(context, e.getMessage(), HttpStatus.BAD_REQUEST);
+            map.put(DatasetWorkingMapKeys.ASSET_NAME_COLLISION, true);
+            return new StepResult(StepStatus.STEP_RESULT_FAILURE_FATAL, e);
+        }
+
         map.put(JobMapKeys.STATUS_CODE.getKeyName(), HttpStatus.CREATED);
         return StepResult.getStepResultSuccess();
     }
 
     @Override
     public StepResult undoStep(FlightContext context) {
-        Dataset dataset = getDataset(context);
-        // Search the Asset list in the dataset object to see if the asset you were trying to create got created.
-        AssetSpecification newAssetSpecification = getNewAssetSpec(context, dataset);
-        Optional<AssetSpecification> assetSpecificationToDelete =
-            dataset.getAssetSpecificationByName(newAssetSpecification.getName());
-        // This only works if we are sure asset names are unique.
-        // You cannot assume that the flight object created when the doStep was run is the same flight object
-        // when the undoStep is run.
-        if (assetSpecificationToDelete.isPresent()) {
-            // If the asset is found, then you get its id and call delete.
-            assetDao.delete(assetSpecificationToDelete.get().getId());
+        FlightMap map = context.getWorkingMap();
+        if (map.get(DatasetWorkingMapKeys.ASSET_NAME_COLLISION, Boolean.class) == null) {
+            Dataset dataset = getDataset(context);
+            // Search the Asset list in the dataset object to see if the asset you were trying to create got created.
+            AssetSpecification newAssetSpecification = getNewAssetSpec(context, dataset);
+            Optional<AssetSpecification> assetSpecificationToDelete =
+                dataset.getAssetSpecificationByName(newAssetSpecification.getName());
+            // This only works if we are sure asset names are unique.
+            // You cannot assume that the flight object created when the doStep was run is the same flight object
+            // when the undoStep is run.
+            if (assetSpecificationToDelete.isPresent()) {
+                // If the asset is found, then you get its id and call delete.
+                assetDao.delete(assetSpecificationToDelete.get().getId());
+            }
         }
         // Else, if the asset is not found, then you are done. It never got created.
         return StepResult.getStepResultSuccess();
